requirements.txt
altgraph==0.17.2
click==8.1.3
colorama==0.4.5
Flask==2.2.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
mss==6.1.0
pefile==2022.5.30
Pillow==9.2.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.8
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
pywin32-ctypes==0.2.0
SQLAlchemy==1.4.40
Werkzeug==2.2.1
